Feature 1 (#1)
* Feature 1

+++ src/main/java/com/example/sliepniev/service/CalculatorService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class CalculatorService {
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
 }
